Maintain entity order after update_inserters
Writing to LuaEntity.direction changes the order
of the entity within the Factorio spatial database,
which changes the order the entity is returned
from LuaSurface.find_entities_filtered().  The
modified entity goes to the front of the list,
so if we process entities in order, the last entity
ends up at the front of the list after iteration.

Since miniloader currently depends on that order
being stable, we need to write to direction in the
reverse of the entities' current order, so
afterwards the original order is restored.

Ideally we should rewrite to not depend on
a stable order, since this hack may break again
later.